fixed a few more places where merit was stored in ints

+++ avida-core/source/cpu/cHardwareBase.cc
@@ -243,7 +243,8 @@ bool cHardwareBase::Divide_CheckViable(cAvidaContext& ctx, const int parent_size
   
   // Determine the fitness of this organism as compared to its parent...
   if (m_world->GetTestSterilize() && !phenotype.IsInjected()) {
-    const int merit_base = phenotype.CalcSizeMerit();
+    //LZ
+    const double merit_base = phenotype.CalcSizeMerit();
     const double cur_fitness = merit_base * phenotype.GetCurBonus() / phenotype.GetTimeUsed();
     const double fitness_ratio = cur_fitness / phenotype.GetLastFitness();
     const Apto::Array<int>& childtasks = phenotype.GetCurTaskCount();
